util: split out sorting related calls to new sort-util.[ch]
