+++ src/libsystemd-network/sd-lldp.c
@@ -15,6 +15,7 @@
 #include "lldp-network.h"
 #include "memory-util.h"
 #include "socket-util.h"
+#include "sort-util.h"
 #include "string-table.h"
 
 #define LLDP_DEFAULT_NEIGHBORS_MAX 128U
